Extensions
Extensions are like plugins but have no arguments and can not interact with the tg api. You can implement a lightweight markov chain or statistics plugin this way. They receive every message sent to the bot + a persistent variable. The persistent variable will initially be None but anything the extension returns will replace that. For example you can keep a DB connection open this way.